firmware: move fpga selftest to a separate file to resolve blinky linker issues
