+++ firmware/common/fpga.h
@@ -23,10 +23,17 @@
 #define __FPGA_H
 
 #include <stdbool.h>
+#include "hackrf_core.h"
 
 /* Up to 5 registers, each containing up to 8 bits of data */
 #define FPGA_NUM_REGS            5
 #define FPGA_DATA_REGS_MAX_VALUE 255
+
+typedef enum {
+	FPGA_QUARTER_SHIFT_MODE_NONE = 0b00,
+	FPGA_QUARTER_SHIFT_MODE_UP = 0b11,
+	FPGA_QUARTER_SHIFT_MODE_DOWN = 0b01,
+} fpga_quarter_shift_mode_t;
 
 struct fpga_driver_t;
 typedef struct fpga_driver_t fpga_driver_t;
@@ -36,6 +43,12 @@ struct fpga_driver_t {
 	uint8_t regs[FPGA_NUM_REGS];
 	uint8_t regs_dirty;
 };
+
+/* Initialize the loaded bitstream's registers to their default values. */
+extern void fpga_init(fpga_driver_t* const drv);
+
+/* Initialize fpga and gateware. */
+extern void fpga_setup(fpga_driver_t* const drv);
 
 /* Read a register via SPI. Save a copy to memory and return
  * value. Mark clean. */
@@ -50,7 +63,17 @@ extern void fpga_reg_write(fpga_driver_t* const drv, uint8_t r, uint8_t v);
  * provided routines for those operations. */
 extern void fpga_regs_commit(fpga_driver_t* const drv);
 
-void fpga_hw_sync_enable(const hw_sync_mode_t hw_sync_mode);
+void fpga_set_hw_sync_enable(fpga_driver_t* const drv, const hw_sync_mode_t hw_sync_mode);
+void fpga_set_rx_dc_block_enable(fpga_driver_t* const drv, const bool enable);
+void fpga_set_rx_decimation_ratio(fpga_driver_t* const drv, const uint8_t value);
+void fpga_set_rx_quarter_shift_mode(
+	fpga_driver_t* const drv,
+	const fpga_quarter_shift_mode_t mode);
+void fpga_set_tx_interpolation_ratio(fpga_driver_t* const drv, const uint8_t value);
+
+void fpga_set_prbs_enable(fpga_driver_t* const drv, const bool enable);
+void fpga_set_tx_nco_enable(fpga_driver_t* const drv, const bool enable);
+void fpga_set_tx_nco_pstep(fpga_driver_t* const drv, const uint8_t phase_increment);
 
 bool fpga_image_load(unsigned int index);
 bool fpga_spi_selftest();
